[refactor] :truck: move getUser to new mongoService (#276)
