+++ src/controllers/balanceController.ts
@@ -2,8 +2,9 @@ import { FastifyReply, FastifyRequest, FastifyInstance } from 'fastify';
 
 import { getPhoneNFTs } from './nftController';
 import { Logger } from '../helpers/loggerHelper';
+import { getUser } from '../services/mongoService';
 import { IUser, IUserWallet } from '../models/user';
-import { getUser, getUserWalletByChainId } from '../services/userService';
+import { getUserWalletByChainId } from '../services/userService';
 import { fetchExternalDeposits } from '../services/externalDepositsService';
 import { isValidPhoneNumber, isValidEthereumWallet } from '../helpers/validationHelper';
 import {
