+++ src/controllers/walletController.ts
@@ -1,11 +1,11 @@
 import { FastifyReply, FastifyRequest } from 'fastify';
 
 import { Logger } from '../helpers/loggerHelper';
+import { getUser } from '../services/mongoService';
 import { IUser, IUserWallet } from '../models/user';
 import { isValidPhoneNumber } from '../helpers/validationHelper';
 import { returnErrorResponse, returnSuccessResponse } from '../helpers/requestHelper';
 import {
-  getUser,
   addWalletToUser,
   createUserWithWallet,
   getUserWalletByChainId
