+++ src/services/notificationService.ts
@@ -3,13 +3,13 @@ import { ethers } from 'ethers';
 import NodeCache from 'node-cache';
 import { channels as PushAPIChannels, payloads as PushAPIPayloads } from '@pushprotocol/restapi';
 
+import { getUser } from './mongoService';
 import { Logger } from '../helpers/loggerHelper';
 import { IBlockchain } from '../models/blockchain';
 import { getNetworkConfig } from './networkService';
-import { User, IUser, IUserWallet } from '../models/user';
+import { IUser, IUserWallet } from '../models/user';
 import { getTemplate, templateEnum } from './templateService';
 import { isValidPhoneNumber } from '../helpers/validationHelper';
-import { getPhoneNumberFormatted } from '../helpers/formatHelper';
 import {
   LanguageEnum,
   ITemplateSchema,
@@ -39,18 +39,6 @@ interface OperatorReplyPayload {
 }
 
 const notificationTemplateCache = new NodeCache({ stdTTL: NOTIFICATION_TEMPLATE_CACHE_TTL });
-/**
- * Retrieves a user based on the phone number.
- * This function is internal to avoid circular imports between userService and notificationService.
- * @param {string} phoneNumber - The phone number of the user to search for.
- * @returns {Promise<IUser | null>} The user corresponding to the provided phone number, or null if no user is found.
- */
-const getUserInternal = async (phoneNumber: string): Promise<IUser | null> => {
-  const user: IUser | null = await User.findOne({
-    phone_number: getPhoneNumberFormatted(phoneNumber)
-  });
-  return user;
-};
 
 /**
  * Retrieves the wallet for a specific chain_id from a user's wallet array.
@@ -124,7 +112,7 @@ export async function sendPushNotificaton(
       return true;
     }
 
-    const user: IUser | null = await getUserInternal(channelUserId);
+    const user: IUser | null = await getUser(channelUserId);
     if (!user) {
       Logger.log(
         'sendPushNotificaton',
@@ -194,7 +182,7 @@ export async function sendPushNotificaton(
 export const getUserSettingsLanguage = async (phoneNumber: string): Promise<LanguageEnum> => {
   let language: LanguageEnum = SETTINGS_NOTIFICATION_LANGUAGE_DFAULT as LanguageEnum;
   try {
-    const user: IUser | null = await getUserInternal(phoneNumber);
+    const user: IUser | null = await getUser(phoneNumber);
     if (user && user.settings) {
       const userLanguage = user.settings.notifications.language;
       if (Object.values(LanguageEnum).includes(userLanguage as LanguageEnum)) {
